requirements/prod.txt
-r base.txt

django_settings @ git+https://github.com/gitkwarteng/django-settings-class.git@0.2.16

components @ git+https://github.com/gitkwarteng/laces-components.git@0.1.1